Add allow wrapping the return result in Identity.
Why:

Often two very similar types exist in the application.
1) The type returned from the server.
2) The type used by a component.
Frequently enough the only difference between the two is that some fields are wrapped in a functor. So if we allow wrapping the results in Identity, we will be able to share the type definition. E.g.

type Data f = { id: ID, name: f String }

type Return = Data Identity
type State = Data (Either Union)

Thus Identity fulfills a similar role to ErrorBoundary, but is a newtype.